backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
chromadb
sentence-transformers
python-dotenv
jinja2
pytest
flake8
google-generativeai

